Slit my plan view from owner logic.
